[FIX]: Updated data types and fixed linting

+++ mobile-wrapper/android/app/src/main/java/com/anonymous/mobilewrapper/models/SpO2Payload.kt
@@ -3,7 +3,7 @@ import com.anonymous.mobilewrapper.models.Range
 
 data class SpO2Point(
     val time: String,
-    val percentage: Double
+    val percentage: Int
 )
 
 data class SpO2Payload(
